Moving dashboard controller into a behavior so it can easily be overridden
